server/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-migrate
flask-bcrypt
pymysql
python-dotenv
marshmallow
marshmallow-sqlalchemy
waitress
stripe
googlemaps
flask-socketio
eventlet
